% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trydustevents-package.R
\docType{package}
\name{trydustevents-package}
\alias{trydustevents}
\alias{trydustevents-package}
\title{trydustevents: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
